requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.7
python-dotenv==1.0.1
jinja2==3.1.3